% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy measures for a forecast model}
\usage{
accuracy(object,...)
}
\value{
A matrix with forecast accuracy measures.
}
\description{
See \code{\link[forecast]{accuracy}} for details.
}
\seealso{
\code{\link[forecast]{accuracy}}
}
\keyword{internal}
