% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autolayer}
\alias{autolayer}
\title{forecast package autolayer function}
\usage{
autolayer(object,...)
}
\value{
A ggplot layer
}
\description{
See \code{\link[forecast]{autolayer}} for details.
}
\seealso{
\code{\link[forecast]{autolayer}}
}
\keyword{internal}
