% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autoplot}
\alias{autoplot}
\title{forecast package autoplot function}
\usage{
autoplot(object,...)
}
\value{
A ggplot object
}
\description{
See \code{\link[forecast]{autoplot}} for details.
}
\seealso{
\code{\link[forecast]{autoplot}}
}
\keyword{internal}
