% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_model.R
\name{CARET}
\alias{CARET}
\title{Tidy time series forecasting using various Machine Learning models.}
\usage{
CARET(
  formula,
  caret_method = "cubist",
  metric = "RMSE",
  initial_window = NULL,
  cv_horizon = 6,
  fixed_window = TRUE,
  verbose = TRUE,
  allow_parallel = FALSE,
  tune_grid = NULL,
  ...
)
}
\arguments{
\item{formula}{A formula object to specify the model (see "Specials" section).}

\item{caret_method}{A string specifying which regression model to use.
Possible values are found using names \code{\link[caret]{getModelInfo}}.
A list of functions can also be passed for a custom model function.
See \url{http://topepo.github.io/caret/} for details.}

\item{metric}{A string that specifies what summary metric will be used to
select the optimal model. See \code{\link[caret]{train}}.}

\item{initial_window}{The initial number of consecutive values in each
training set sample. Please set to NULL, if your data is a unbalanced panel
data.}

\item{cv_horizon}{The number of consecutive values in test set sample.}

\item{fixed_window}{Logical, if FALSE, all training samples start at 1.}

\item{verbose}{A logical for printing a training log.}

\item{allow_parallel}{If a parallel backend is loaded and available,
should the function use it?}

\item{tune_grid}{A data frame with possible tuning values.
The columns are named the same as the tuning parameters.
Use getModelInfo to get a list of tuning parameters for each model or
see \url{http://topepo.github.io/caret/available-models.html}.
(NOTE: If given, this argument must be named.)}

\item{...}{Ignored.}
}
\value{
For Formula interface:

A model specification.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Tidy time series forecasting using fabletools and caret packages.
}
\section{Specials}{


\subsection{order}{
The \code{order} special is used to specify the lag order for the auto-regression.
\preformatted{
order(p = 1, period = NULL)
}
\itemize{
\item \code{p} he order of the auto-regressive (AR) terms. Only one integer allowed
\item \code{period} The periodic nature of the seasonality. This can be either a
number indicating the number of observations in each seasonal period, or
text to indicate the duration of the seasonal window (for example, annual
seasonality would be "1 year").
}

}

\subsection{xreg}{
Exogenous regressors can be included in an AR model without explicitly using
the \code{xreg()} special. Common exogenous regressor specials as specified in
\code{\link{common_xregs}} can also be used. These regressors are handled using
\code{\link[stats:model.frame]{stats::model.frame()}}, and so interactions and other functionality
behaves similarly to \code{\link[stats:lm]{stats::lm()}}.

The inclusion of a constant in the model follows the similar rules to
\code{\link[stats:lm]{stats::lm()}}, where including \code{1} will add a constant and \code{0} or \code{-1}
will remove the constant. If left out, the inclusion of a constant will be
determined by minimising \code{ic}.

\preformatted{
xreg(..., fixed = list())
}
\itemize{
\item \code{...} are expressions for the exogenous regressors (such as \code{log(x)})
\item \code{fixed} A named list of fixed parameters for coefficients.
The names identify the coefficient, and should match the name of the
regressor.
For example, \code{fixed = list(constant = 20)}.
}

}
}

\examples{



model_spec = CARET(
  object = value ~ order(12) + trend() + fourier(K = 6),
  caret_method = "lm",
  verbose = FALSE)

AirPassengers \%>\% tsibble::as_tsibble() \%>\%
  model(model_spec) -> fit

fc_df <- forecast(fit, h = 23)

fc_df\%>\%autoplot(AirPassengers\%>\% tsibble::as_tsibble())

}
