% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretForecast-package.R
\docType{package}
\name{caretForecast-package}
\alias{caretForecast}
\alias{caretForecast-package}
\title{caretForecast: Time Series Forecasting Using Caret Infrastructure}
\description{
Recursive time series forecast using Caret infrastructure. The models are selected based on time series cross-validation and forecasting is done recursively.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Akai01/caretForecast}
  \item Report bugs at \url{https://github.com/Akai01/caretForecast/issues}
}

}
\author{
\strong{Maintainer}: Resul Akay \email{resulakay1@gmail.com}

}
\keyword{internal}
