% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{fc_plot}
\alias{fc_plot}
\title{Plot a forecast or forecastARml object}
\usage{
fc_plot(object, ...)
}
\arguments{
\item{object}{a forecast object}

\item{...}{other arguments passed to \code{\link[forecast]{autoplot}}}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\author{
Res Akay
}
