% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_model.R
\name{fitted.CARET}
\alias{fitted.CARET}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{CARET}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values.
}
\examples{

library(tsibble)
USAccDeaths \%>\%
  as_tsibble() \%>\%
  model(caret = CARET(log(value) ~ order(5))) \%>\%
  fitted()

}
