% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_model.R
\name{forecast.CARET}
\alias{forecast.CARET}
\title{Forecast a CARET model}
\usage{
\method{forecast}{CARET}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{new_data}{A tsibble containing future information used to forecast.}

\item{specials}{Model specials, passed by fabletools::forecast.mdl_df().}

\item{...}{Additional arguments for forecast model methods.}
}
\description{
Forecast a CARET model
}
