% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forecast_loop_cal}
\alias{forecast_loop_cal}
\title{Internal forecast loop for calibration (without time series attributes)}
\usage{
forecast_loop_cal(object, xreg, h)
}
\description{
Internal forecast loop for calibration (without time series attributes)
}
\keyword{internal}
