% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_conformal.R
\name{predict.conformalRegressor}
\alias{predict.conformalRegressor}
\title{Predict a conformalRegressor}
\usage{
\method{predict}{conformalRegressor}(
  object,
  y_hat = NULL,
  sigmas = NULL,
  confidence = 0.95,
  y_min = -Inf,
  y_max = Inf,
  ...
)
}
\arguments{
\item{object}{A conformalRegressor object}

\item{y_hat}{Predicted values}

\item{sigmas}{Difficulty estimates}

\item{confidence}{Confidence level}

\item{y_min}{The minimum value to include in prediction intervals}

\item{y_max}{The maximum value to include in prediction intervals}

\item{...}{Ignored}
}
\value{
Prediction intervals
}
\description{
Predict a conformalRegressor
}
\author{
Resul Akay
}
