% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{retail}
\alias{retail}
\title{Grouped sales data from an Australian Retailer}
\format{
A data class of "tbl_df", "tbl", "data.frame" with 13986 rows and 3 columns:
\describe{
\item{date}{date}
\item{item}{products}
\item{value}{sales}
}
}
\source{
\url{https://robjhyndman.com/data/ausretail.csv}
}
\usage{
retail
}
\description{
A dataset containing 42 products' sales
}
\keyword{datasets}
