% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{retail_wide}
\alias{retail_wide}
\title{Sales data from an Australian Retailer in time series format}
\format{
An object of class mts (inherits from ts, matrix)
with 333 rows and 43 columns.
\describe{
This data set is the wide format of \code{\link{retail}} data.
}
}
\source{
\url{https://robjhyndman.com/data/ausretail.csv}
}
\usage{
retail_wide
}
\description{
A dataset containing 42 products' sales
}
\keyword{datasets}
