% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_ts.R
\name{split_ts}
\alias{split_ts}
\title{Split a time series into training and testing sets}
\usage{
split_ts(y, test_size = 10)
}
\arguments{
\item{y}{Time series}

\item{test_size}{The number of observations to keep in the test set}
}
\description{
Split a time series into training and testing sets
}
\author{
Resul Akay
}
