% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested_methods.R
\name{suggested_methods}
\alias{suggested_methods}
\title{Suggested methods for ARml}
\usage{
suggested_methods()
}
\value{
A character vector of Suggested methods
}
\description{
Suggested methods for ARml
}
\author{
Res Akay
}
