% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_conformal_reg.R
\name{conformalRegressor}
\alias{conformalRegressor}
\title{Fit a conformal regressor.}
\usage{
conformalRegressor(residuals, sigmas = NULL)
}
\arguments{
\item{residuals}{Model residuals.}

\item{sigmas}{A vector of difficulty estimates}
}
\value{
A conformalRegressor object
}
\description{
Fit a conformal regressor.
}
\references{
Boström, H., 2022. crepes: a Python Package for Generating Conformal
Regressors and Predictive Systems. In Conformal and Probabilistic Prediction
and Applications. PMLR, 179.
}
\author{
Resul Akay
}
