% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{fc_plot}
\alias{fc_plot}
\title{Plot a forecast or forecastARml object}
\usage{
fc_plot(object, ...)
}
\arguments{
\item{object}{a forecast object}

\item{...}{other arguments passed to \code{\link[forecast]{autoplot}}}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{
\dontrun{
library(forecast)
library(caretForecast)
train <- window(AirPassengers, end = c(1959, 12))

test <- window(AirPassengers, start = c(1960, 1))

ARml(train, caret_method = "svmLinear2", maxlag = 12, trend_method = "none",
 pre_process = "center") -> fit

forecast(fit, h = length(test), level = NULL, PI = T) -> fc
fc_plot(fc)+ autolayer(test)
accuracy(fc, test)
}

}
\author{
Resul Akay
}
