% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_imp.R
\name{get_var_imp}
\alias{get_var_imp}
\title{Calculation of variable importance for forecasting model.}
\usage{
get_var_imp(object, plot = T)
}
\arguments{
\item{object}{A list class of ARml or forecast object derived from ARml}

\item{plot}{Boolean, if TRUE, variable importance will be ploted.}
}
\value{
A list class of "varImp.train". See \code{\link[caret]{varImp}}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{
\dontrun{
library(MLforecast)
library(forecast)

train <- window(AirPassengers, end = c(1959, 12))

test <- window(AirPassengers, start = c(1960, 1))

ARml(train, caret_method = "svmLinear2", maxlag = 12, trend_method = "none",
 pre_process = "center") -> fit

forecast(fit, h = length(test), level = NULL, PI = T) -> fc

plot(fc)+ autolayer(test)

accuracy(fc, test)

get_var_imp(fc, plot = T)
}


}
\author{
Res Akay
}
