% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_ts}
\alias{split_ts}
\title{Split a time series into training and testing sets}
\usage{
split_ts(y, test_size = 10)
}
\arguments{
\item{y}{A univariate time series}

\item{test_size}{The number of observations to keep in the test set}
}
\value{
A list with train and test elements
}
\description{
Split a time series into training and testing sets
}
\examples{


dlist <- split_ts(retail_wide[,1], test_size = 12)




}
\author{
Resul Akay
}
