% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_model.R, R/utils.R
\name{suggested_methods}
\alias{suggested_methods}
\title{Suggested methods for CARET}
\usage{
suggested_methods()

suggested_methods()
}
\value{
A character vector of Suggested methods

A character vector of Suggested methods
}
\description{
Suggested methods for CARET

Suggested methods for ARml
}
\examples{

\dontrun{

suggested_methods()

}


suggested_methods()

}
\author{
Resul Akay

Resul Akay
}
