% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{ts_plot}
\alias{ts_plot}
\title{Plot a ts object}
\usage{
ts_plot(object, ...)
}
\arguments{
\item{object}{a ts object}

\item{...}{other arguments}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\author{
Res Akay
}
